backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
PyPDF2==3.0.1
Werkzeug==3.1.3
python-pptx==1.0.2
fuzzywuzzy==0.18.0
sentence_transformers